notebooks/requirements.txt
httpx>=0.27
psycopg2-binary>=2.9
matplotlib>=3.9
numpy>=1.26
jupyter>=1.0
ipykernel>=6.29
anthropic>=0.40
opencv-python-headless>=4.9
Pillow>=10.0
svgpathtools>=1.6
vtracer>=0.6
